+++ src/main/java/org/folio/rest/converter/common/attr/ProxyConverter.java
@@ -1,22 +1,22 @@
 package org.folio.rest.converter.common.attr;
 
 import java.util.Objects;
-import org.folio.holdingsiq.model.Proxy;
+import org.folio.holdingsiq.model.ProxyUrl;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ProxyConverter implements Converter<Proxy, org.folio.rest.jaxrs.model.Proxy> {
-
+public class ProxyConverter implements Converter<ProxyUrl, org.folio.rest.jaxrs.model.ProxyUrl> {
   @Override
-  public org.folio.rest.jaxrs.model.Proxy convert(@Nullable Proxy proxy) {
+  public org.folio.rest.jaxrs.model.ProxyUrl convert(@Nullable ProxyUrl proxy) {
     if (Objects.isNull(proxy)) {
       return null;
     }
-    org.folio.rest.jaxrs.model.Proxy p = new org.folio.rest.jaxrs.model.Proxy();
+    org.folio.rest.jaxrs.model.ProxyUrl p = new org.folio.rest.jaxrs.model.ProxyUrl();
     p.setId(proxy.getId());
     p.setInherited(proxy.getInherited());
+    p.setProxiedUrl(proxy.getProxiedUrl());
     return p;
   }