+++ src/test/java/org/folio/rest/impl/integrationsuite/EholdingsResourcesImplTest.java
@@ -387,6 +387,23 @@ public class EholdingsResourcesImplTest extends WireMockTestBase {
   }
 
   @Test
+  public void shouldAcceptValuesCustomResourceWithUnrecognizedFieldInProxy()
+    throws IOException, URISyntaxException, JSONException {
+    String stubResponseFile = "responses/rmapi/resources/get-custom-resource-updated-response.json";
+    String expectedResourceFile = "responses/kb-ebsco/resources/expected-custom-resource.json";
+
+    String actualResponse =
+      mockUpdateResourceScenario(stubResponseFile, CUSTOM_RESOURCE_ENDPOINT, STUB_CUSTOM_RESOURCE_ID,
+        readFile("requests/kb-ebsco/resource/put-custom-resource-with-proxied-url.json"));
+
+    JSONAssert.assertEquals(readFile(expectedResourceFile), actualResponse, false);
+
+    verify(1, putRequestedFor(new UrlPathPattern(new RegexPattern(CUSTOM_RESOURCE_ENDPOINT), true))
+      .withRequestBody(
+        equalToJson(readFile("requests/rmapi/resources/put-custom-resource-is-selected-multiple-attributes.json"))));
+  }
+
+  @Test
   public void shouldUpdateTagsOnSuccessfulTagsPut() throws IOException, URISyntaxException {
     List<String> tags = Collections.singletonList(STUB_TAG_VALUE);
     sendPutTags(tags);